refactor(mountains-detail): 탭 컴포넌트 features/mountains-detail로 분리
- CourseTab, TransportTab, AmenityTab, RestaurantTab, ReviewTab 컴포넌트 분리
- CourseCard props를 course 객체 대신 개별 필드로 변경
- API CourseInfo 타입 기준으로 필드명 통일 (courseId, name, distance, duration)
