+++ app/mountains/[id].tsx
@@ -1,30 +1,17 @@
-import { BusIcon } from "@/components/icons/bus-icon";
-import { CarIcon } from "@/components/icons/car-icon";
 import { CaretDownIcon } from "@/components/icons/caret-down-icon";
 import { CaretLeftIcon } from "@/components/icons/caret-left-icon";
 import { HeartIcon } from "@/components/icons/heart-icon";
-import { InfoCenterIcon } from "@/components/icons/info-center-icon";
-import { MagicWandIcon } from "@/components/icons/magic-wand-icon";
-import { MegaphoneIcon } from "@/components/icons/megaphone-icon";
-import { ParkingIcon } from "@/components/icons/parking-icon";
-import { ShelterIcon } from "@/components/icons/shelter-icon";
-import { StoreIcon } from "@/components/icons/store-icon";
-import { SubwayIcon } from "@/components/icons/subway-icon";
 import { SunriseIcon } from "@/components/icons/sunrise-icon";
 import { SunsetIcon } from "@/components/icons/sunset-icon";
-import { ToiletIcon } from "@/components/icons/toilet-icon";
-import { UserIcon } from "@/components/icons/user-icon";
-import { CourseBadge } from "@/features/mountains/components/course-badge";
+import { AmenityTab } from "@/features/mountains-detail/components/amenity-tab";
+import { CourseTab } from "@/features/mountains-detail/components/course-tab";
+import { RestaurantTab } from "@/features/mountains-detail/components/restaurant-tab";
+import { ReviewTab } from "@/features/mountains-detail/components/review-tab";
+import { TransportTab } from "@/features/mountains-detail/components/transport-tab";
 import {
   DIFFICULTY_LABEL,
   DIFFICULTY_STYLE,
 } from "@/features/mountains/components/mountain-card";
-import {
-  MOCK_COURSES,
-  MOCK_REVIEWS,
-  RESTAURANT_SECTIONS,
-  type Course,
-} from "@/features/mountains/constants/mountain-detail";
 import { useMountainDetail } from "@/features/mountains/hooks/use-mountain-detail";
 import { buildWeatherDays } from "@/features/mountains/modules/weather";
 import { LinearGradient } from "expo-linear-gradient";
@@ -65,297 +52,6 @@ function SunriseSunset({
           <Text className="text-label-subtler typo-body-3-medium">일몰</Text>
         </View>
         <Text className="text-label-subtle typo-body-3-medium">{sunset}</Text>
-      </View>
-    </View>
-  );
-}
-
-function CourseCard({ course }: { course: Course }) {
-  const router = useRouter();
-  return (
-    <TouchableOpacity
-      className="flex-row items-center gap-4"
-      activeOpacity={0.7}
-      onPress={() => router.push(`/mountains/courses/${course.id}`)}
-    >
-      <View className="h-[72px] w-16 overflow-hidden rounded-[10px] bg-fill-stronger">
-        <Image
-          source={course.imageSource ?? { uri: course.imageUrl }}
-          style={{ width: "100%", height: "100%" }}
-          resizeMode="stretch"
-        />
-      </View>
-      <View className="gap-1.5">
-        <View className="flex-row items-center gap-1.5">
-          <CourseBadge difficulty={course.difficulty} />
-          <Text className="text-label-normal typo-body-1-normal-semi-bold">
-            {course.title}
-          </Text>
-        </View>
-        <View className="flex-row items-center gap-1.5">
-          <Text className="text-label-subtler typo-caption-1-medium">
-            {course.distanceKm}km
-          </Text>
-          <View className="size-[2px] rounded-full bg-label-subtler" />
-          <Text className="text-label-subtler typo-caption-1-medium">
-            {course.durationHours}시간
-          </Text>
-        </View>
-      </View>
-    </TouchableOpacity>
-  );
-}
-
-// ── 교통 ──────────────────────────────────────────────
-type TransportItem = {
-  id: number;
-  title: string;
-  description: string;
-  icon?: React.JSX.Element;
-};
-type TransportSection = {
-  heading: string;
-  items: TransportItem[];
-};
-const TRANSPORT_SECTIONS: TransportSection[] = [
-  {
-    heading: "대중교통",
-    items: [
-      {
-        id: 1,
-        title: "지하철",
-        description: "2호선 서울대입구역 하차 후 5511번 버스 탑승",
-        icon: <SubwayIcon size={16} color="#5C6170" />,
-      },
-      {
-        id: 2,
-        title: "버스",
-        description: "5511, 5513번 버스 이용 (관악산 입구 하차)",
-        icon: <BusIcon size={16} color="#5C6170" />,
-      },
-    ],
-  },
-  {
-    heading: "주차장",
-    items: [
-      {
-        id: 1,
-        title: "서울대 정문",
-        description: "서울대학교 정문 주차장 (유료, 30분 1,000원)",
-        icon: <CarIcon size={16} color="#5C6170" />,
-      },
-      {
-        id: 2,
-        title: "과천 향교 방면",
-        description: "과천 방면 공영주차장 (무료, 100대 수용)",
-        icon: <CarIcon size={16} color="#5C6170" />,
-      },
-    ],
-  },
-];
-
-// ── 편의시설 ────────────────────────────────────────────
-type Facility = "화장실" | "안내소" | "쉼터" | "주차장" | "매점";
-type AmenityArea = {
-  name: string;
-  facilities: Facility[];
-};
-const AMENITY_AREAS: AmenityArea[] = [
-  {
-    name: "서울대 입구",
-    facilities: ["화장실", "안내소", "쉼터", "주차장", "매점"],
-  },
-  { name: "신림 방향", facilities: ["화장실", "쉼터", "매점"] },
-  { name: "과천 방향", facilities: ["화장실", "안내소", "주차장"] },
-];
-
-// ── 맛집 / 리뷰 ── constants/mountain-detail.ts 참고
-
-// ── 탭 컴포넌트 ──────────────────────────────────────────
-function TransportTab() {
-  return (
-    <View className="w-full gap-10 px-6">
-      {TRANSPORT_SECTIONS.map((section) => (
-        <View key={section.heading} className="gap-4">
-          <Text className="text-label-normal typo-headline-1-semi-bold">
-            {section.heading}
-          </Text>
-          <View className="gap-4">
-            {section.items.map((item) => (
-              <View key={item.id} className="flex-row items-start gap-3">
-                <View className="items-center justify-center rounded-full bg-fill-stronger p-1">
-                  {item.icon ?? (
-                    <View className="size-5 rounded-[4px] bg-fill-stronger" />
-                  )}
-                </View>
-                <View className="flex-1 gap-2">
-                  <Text className="text-label-normal typo-body-1-normal-semi-bold">
-                    {item.title}
-                  </Text>
-                  <Text className="text-label-subtle typo-body-2-reading-regular">
-                    {item.description}
-                  </Text>
-                </View>
-              </View>
-            ))}
-          </View>
-        </View>
-      ))}
-    </View>
-  );
-}
-
-const FACILITY_ICON: Record<Facility, React.JSX.Element> = {
-  화장실: <ToiletIcon />,
-  안내소: <InfoCenterIcon />,
-  쉼터: <ShelterIcon />,
-  주차장: <ParkingIcon />,
-  매점: <StoreIcon />,
-};
-
-function AmenityTab() {
-  return (
-    <View className="w-full px-5">
-      <Text className="mb-4 text-label-normal typo-headline-1-semi-bold">
-        주요 편의시설
-      </Text>
-      <View className="gap-8">
-        {AMENITY_AREAS.map((area) => (
-          <View key={area.name} className="gap-3">
-            <Text className="text-label-subtle typo-body-1-normal-semi-bold">
-              {area.name}
-            </Text>
-            <View className="flex-row flex-wrap">
-              {area.facilities.map((facility) => (
-                <View key={facility} className="items-center gap-1.5 p-4">
-                  {FACILITY_ICON[facility]}
-                  <Text className="text-center text-label-subtle typo-body-3-regular">
-                    {facility}
-                  </Text>
-                </View>
-              ))}
-            </View>
-          </View>
-        ))}
-      </View>
-    </View>
-  );
-}
-
-function RestaurantTab() {
-  return (
-    <View className="w-full gap-8">
-      {RESTAURANT_SECTIONS.map((section) => (
-        <View key={section.title} className="gap-4">
-          <Text className="px-5 text-label-normal typo-headline-1-semi-bold">
-            {section.title}
-          </Text>
-          <ScrollView
-            horizontal
-            showsHorizontalScrollIndicator={false}
-            contentContainerStyle={{ paddingHorizontal: 24, gap: 12 }}
-          >
-            {section.items.map((item) => (
-              <View key={item.id} className="gap-2">
-                <View className="h-[116px] w-[188px] rounded-[10px] bg-fill-stronger" />
-                <View className="gap-0.5">
-                  <Text className="text-label-normal typo-body-1-normal-semi-bold">
-                    {item.name}
-                  </Text>
-                  <Text className="text-label-subtler typo-caption-1-medium">
-                    {item.category}
-                  </Text>
-                </View>
-              </View>
-            ))}
-            <View className="h-[116px] w-[188px] items-center justify-center gap-1 rounded-[10px] bg-fill-stronger">
-              <Text className="text-center text-label-subtle typo-body-2-normal-semi-bold">
-                {section.moreLabel}
-              </Text>
-              <Text className="text-label-subtler typo-body-2-normal-regular">
-                {"더보기 >"}
-              </Text>
-            </View>
-          </ScrollView>
-        </View>
-      ))}
-    </View>
-  );
-}
-
-function ReviewTab() {
-  return (
-    <View className="w-full gap-8 px-5">
-      {/* 요약 카드 */}
-      <View className="gap-3">
-        <View className="gap-2 rounded-[12px] bg-blue-25 px-5 py-[18px]">
-          <View className="flex-row items-center gap-2">
-            <MagicWandIcon />
-            <Text className="text-label-normal typo-body-1-normal-semi-bold">
-              한줄 경험 요약
-            </Text>
-          </View>
-          <Text className="text-label-subtle typo-body-2-normal-regular">
-            생각보다 더 힘들고, 그만큼 정상에서의 보람이 큰 산
-          </Text>
-        </View>
-        <View className="gap-2 rounded-[12px] bg-red-25 px-5 py-[18px]">
-          <View className="flex-row items-center gap-2">
-            <MegaphoneIcon />
-            <Text className="text-label-normal typo-body-1-normal-semi-bold">
-              가기 전에 꼭 알아야 할 한 가지
-            </Text>
-          </View>
-          <Text className="text-label-subtle typo-body-2-normal-regular">
-            중간에 화장실 거의 없음, 초입에서 미리 다녀오세요
-          </Text>
-        </View>
-      </View>
-
-      {/* 커뮤니티 리뷰 */}
-      <View className="gap-4">
-        <View className="flex-row items-center gap-2">
-          <Text className="text-label-normal typo-headline-1-semi-bold">
-            커뮤니티 리뷰
-          </Text>
-          <Text className="text-neutral-300 typo-headline-1-semi-bold">54</Text>
-        </View>
-
-        <View className="gap-4">
-          {MOCK_REVIEWS.map((review) => (
-            <View key={review.id} className="gap-3 px-1 pb-4 pt-1">
-              <View className="flex-row gap-4">
-                <View className="size-[92px] rounded-[10px] bg-fill-stronger" />
-                <View className="flex-1 gap-1.5">
-                  <View className="flex-row items-center gap-1.5">
-                    <View className="items-center justify-center rounded-full bg-fill-strongest p-[3px]">
-                      <UserIcon />
-                    </View>
-                    <Text className="text-label-normal typo-body-2-normal-semi-bold">
-                      {review.userName}
-                    </Text>
-                  </View>
-                  <Text
-                    className="text-label-normal typo-body-2-normal-regular"
-                    numberOfLines={3}
-                  >
-                    {review.text}
-                  </Text>
-                </View>
-              </View>
-              <View className="flex-row items-center gap-2">
-                <CourseBadge difficulty={review.difficulty} />
-                <Text className="text-label-normal typo-body-2-normal-medium">
-                  {review.courseName}
-                </Text>
-              </View>
-            </View>
-          ))}
-        </View>
-
-        <TouchableOpacity className="h-[38px] items-center justify-center rounded-[8px] bg-fill-stronger">
-          <Text className="text-label-subtle typo-label-medium">더보기</Text>
-        </TouchableOpacity>
       </View>
     </View>
   );
@@ -505,13 +201,7 @@ export default function MountainDetailScreen() {
               </View>
 
               {/* Tab content */}
-              {activeTab === "코스" && (
-                <View className="w-full gap-4 px-5">
-                  {MOCK_COURSES.map((course) => (
-                    <CourseCard key={course.id} course={course} />
-                  ))}
-                </View>
-              )}
+              {activeTab === "코스" && <CourseTab />}
               {activeTab === "교통" && <TransportTab />}
               {activeTab === "편의" && <AmenityTab />}
               {activeTab === "맛집" && <RestaurantTab />}
